% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calcHet}
\alias{calcHet}
\title{Calculate individual heterozygosity}
\usage{
calcHet(gt, nLoci = NULL)
}
\arguments{
\item{gt}{A genotype matrix with \emph{N} rows and
\emph{L} columns, where \emph{N} is the number
of individuals and \emph{L} is the number of
loci, for which the \emph{i},\emph{j}th element
gives the number of the counted allele in each
individual at each locus.}

\item{nLoci}{An optional argument giving the number of genotyped
base pairs in each individual. If not specified,
this is assumed to be the number of non-missing
genotypes in each row of \code{gt}. Default is
\code{NULL}.}
}
\value{
A vector giving the proportion of loci that are
heterozygous in each individual.
}
\description{
Calculate individual heterozygosity
}
\details{
The data matrix \code{gt} should consist of
0s, 1s, and 2s, with missing data indicated with
\code{NA}. This can be generated from a VCF file
using \code{\link{vcf2R}}. If \code{nLoci} is not
specified, and \code{gt} consists only of
polymorphic loci, this function will return
heterozygosity at polymorphic sites, which may not
be comparable across datasets.
}
