% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calcPWP}
\alias{calcPWP}
\title{Calculate pairwise pi between two diploid individuals}
\usage{
calcPWP(ind1, ind2, L)
}
\arguments{
\item{ind1}{A genotype vector for a single diploid individual
for which the \emph{i}th element
gives the frequency of the counted
allele in that individual at the \emph{i}th locus.}

\item{ind2}{A genotype vector for a single diploid individual
for which the \emph{i}th element
gives the frequency of the counted
allele in that individual at the \emph{i}th locus.}

\item{L}{The number of loci genotyped in \emph{both} individuals 1
and 2.}
}
\value{
An estimate of pairwise pi between individuals 1 and 2.
}
\description{
Calculate pairwise pi between two diploid individuals
}
\details{
The genotype data should consist of {0,0.5,1}
with missing data indicated with \code{NA}.
If the genotype vectors \code{ind1} and \code{ind2}
consist only of polymorphic loci, and \code{L} is the
number of co-genotyped polymorphic loci,
this function will return pairwise pi at
polymorphic sites, which may not be
comparable across datasets.
}
