% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calcThetaW}
\alias{calcThetaW}
\title{Calculate Wu and Watterson's Theta}
\usage{
calcThetaW(gt, lociDistn = NULL)
}
\arguments{
\item{gt}{A genotype matrix with \emph{N} rows and
\emph{L} columns, where \emph{N} is the number
of individuals and \emph{L} is the number of
loci, for which the \emph{i},\emph{j}th element
gives the number of the counted allele in each
individual at each locus.}

\item{lociDistn}{A vector of length \emph{N},
for which the \emph{i}th element
gives the number of base pairs genotyped in exactly
\emph{i} individuals. E.g., the 3rd element of
\code{lociDistn} gives the number of base pairs
genotyped in exactly 3 individuals. Default is
\code{NULL}.}
}
\value{
An estimate of Wu and Watterson's Theta.
}
\description{
Calculate Wu and Watterson's Theta
}
\details{
The data matrix \code{gt} should consist of
0s, 1s, and 2s, with missing data indicated with
\code{NA}. This can be generated from a VCF file
using \code{\link{vcf2R}}, and the \code{lociDistn}
can be generated using \code{\link{getBPstats}}.
If \code{lociDistn} is not specified,
and \code{gt} consists only of polymorphic
loci, this function will return Wu and
Watterson's Theta at polymorphic sites,
which may not be comparable across datasets.
}
