% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{doAllTheThings}
\alias{doAllTheThings}
\title{Calculate a bunch of popgen stats from a VCF file}
\usage{
doAllTheThings(vcfFile, lociDistn = NULL, coGeno = NULL, nPCs = 4, outPath)
}
\arguments{
\item{vcfFile}{The filename (with necessary path) of the
VCF file you want to read into R.}

\item{lociDistn}{A vector of length \emph{N}, where \emph{N} is the
number of genotyped individuals, for which the
\emph{i}th element gives the number of base pairs
genotyped in exactly \emph{i} individuals.
E.g., the 3rd element of \code{lociDistn} gives
the number of base pairs genotyped in exactly 3
individuals. Default is \code{NULL}.}

\item{coGeno}{A symmetric matrix (dimensions \emph{N} x \emph{N})
for which the \emph{i},\emph{j}th element gives
the number of base pairs genotyped in \emph{both}
samples \emph{i} and \emph{j}. The order of samples in
this matrix must be the same as the order of samples
in the VCF file. If left unspecified, each cell will
be assigned a value of \code{L} (i.e., indicating
no missing data). Default is \code{NULL}.}

\item{nPCs}{The number of principal component axes to retain.
Default is 4.}

\item{outPAth}{The filename (with necessary path) of the
output object you want to generate.}
}
\value{
This function does not return a value. Instead, a
named list is saved as a .Robj file at the location
designated by the \code{outPath} argument. The elements
of the list are:
\itemize{
\item \code{thetaW} Wu and Watterson's theta,
as calculated by \code{\link{calcThetaW}}
\item \code{pwp} a matrix of pairwise pi between
all samples, as calculated by \code{\link{freqs2pairwisePi}}
\item \code{globalPi} global pi in the dataset
(the mean of the upper-triangle of \code{pwp})
\item \code{pcs} the first \code{nPCs} principal
components, as calculated by \code{\link{doPCA}}
\item \code{het} proportion of heterozygous loci in
each individual, as calculated by \code{\link{calcHet}}.
}
}
\description{
Calculate a bunch of popgen stats from a VCF file
}
