% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{doPCA}
\alias{doPCA}
\title{Run a principal components analysis}
\usage{
doPCA(gt, nPCs = 4)
}
\arguments{
\item{gt}{A genotype matrix with \emph{N} rows and
\emph{L} columns, where \emph{N} is the number
of individuals and \emph{L} is the number of
loci, for which the \emph{i},\emph{j}th element
gives the number of the counted allele in each
individual at each locus.}

\item{nPCs}{The number of principal component axes to retain.
Default is 4.}
}
\value{
A matrix with \emph{N} rows and \code{nPCs} columns
giving the position of each of the \emph{N} samples on
the first \code{nPCs} principal component axes.
}
\description{
Run a principal components analysis
}
