% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{freqs2pairwisePi}
\alias{freqs2pairwisePi}
\title{Calculate pairwise pi between all pairs of individuals in a dataset}
\usage{
freqs2pairwisePi(freqs, coGeno = NULL, quiet = FALSE)
}
\arguments{
\item{freqs}{A genotype matrix with \emph{N} rows and
\emph{L} columns, where \emph{N} is the number
of individuals and \emph{L} is the number of
loci, for which the \emph{i},\emph{j}th element
gives the frequency of the counted allele in each
individual at each locus.}

\item{coGeno}{A symmetric matrix (dimensions \emph{N} x \emph{N})
for which the \emph{i},\emph{j}th element gives
the number of base pairs genotyped in \emph{both}
samples \emph{i} and \emph{j}. Default is \code{NULL}.
If left unspecified, each cell will be assigned a value of
\code{L} (i.e., indicating no missing data).}

\item{quiet}{A \code{logical} value that controls whether
a progress par is displayed. Default is \code{FALSE}.}
}
\value{
An estimate of pairwise pi between all individuals
in the data matrix.
}
\description{
Calculate pairwise pi between all pairs of individuals in a dataset
}
\details{
The genotype data should consist of {0,0.5,1}
with missing data indicated with \code{NA}.
If the genotype matrix consists only of
polymorphic loci, and \code{coGeno} either
contains the number of co-genotyped polymorphic
loci for each sample pair or is not specified
this function will return pairwise pi at
polymorphic sites, which may not be
comparable across datasets.
}
