% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{getBPstats}
\alias{getBPstats}
\title{Generate genotyped base-pair statistics}
\usage{
getBPstats(stacksFAfile, outPath)
}
\arguments{
\item{stacksFAfile}{The filename (with necessary path) of the
.fa file output by a STACKS run.}

\item{outPAth}{The filename (with necessary path) of the
output object you want to generate.}
}
\value{
This function does not return a value. Instead, a
named list is saved as a .Robj file at the location
designated by the \code{outPath} argument. The elements
of the list are:
\itemize{
\item \code{lociDistn} vector of length \emph{N},
where \emph{N} is total number of individuals
in the dataset, for with the \emph{i}th element
gives the number of base pairs genotyped in exactly
\emph{i} individuals. E.g., the 3rd element of
\code{lociDistn} gives the number of base pairs
genotyped in exactly 3 individuals.
\item \code{coGeno} a symmetric matrix (dimensions N x N)
for which the \link[=\\emph{i},\\emph{j}th]{\emph{i},\emph{j}th} element gives
the number of base pairs genotyped in \emph{both}
samples \emph{i} and \emph{j}.
}
}
\description{
Generate genotyped base-pair statistics
}
